Massive Update: Using a runtime.js file
In this update things have changed a lot. I have changed to using a
runtime.js file that defines classes and methods for the various whiley
datatypes. This means whenever a datatype is created, the corresponding
datatype will be created from the runtime file. The testing harness has
been updated to use Rhino.js, as node could not run with a runtime file.