Tidy up code base.
The code base was in a bit of a mess, and this commit makes quite a lot
of changes.  In particular, it removes a large number of temporary or
junk files.  I have also removed all the test case files as these should
be obtained from the official whiley test directory.  Have updated the
test harness to successfully compile the whiley files to wyil files.
Have now to generate JavaScript and then execute that.