Add Type.Byte functionality to the methods in WyJS.java issue #5
Methods into WyJS.java that uses WyJS.Byte now may call
the type. One test now is passing for because it can create
the Byte object (While_Valid_18).

Methods write() and getType().

As mostly the syntax in the js file genetared is correct,
the js files form the tests that weren't being able to be
generated, now are being generated.