Fix bitwise operators issue #5
Fix bitwise operators by creating the byte type already
using integers instead of strings representing the byte.

So instead of using WyJS.Byte("0000001b") when creating the
byte type in WyJS.java - because I was passing it as varByte
directly before. Now it's using varByte.value which already
gets the integer representing that byte array.

By doing that, all the problems with converting the bytes
and the types were solved because the operators in JS work
with integers. The operators now work properly.