Fix bitwise invert function issue #5
Fix problem with bitwise invert function in byte code.
The bitwise NOT operator in JavaScript has a signed output
that was causing the tests with the repective operator to fail.
The same solution used with the java part was used, adding `result = result &
(0xff);` to so the result could be converted to unsigned integer.