Terminating "Original Implementation".
This hereby terminates the original version of the Whiley2JavaScript
compiler.  A copy of that version is kept under the branch
"OriginalImplementation" for historial reasons.

The Whiley2JavaScript compiler will now be completely reimplemented from
scratch, potentially under a new license (Apache 2.0).  The starting
point for this rewrite will be the current "Whiley2Java" compiler which
provides much of the necessary foundations.

The main reason for rewriting the Whiley2JavaScript compiler is that the
WhileyCompiler (v0.4.0+) itself has changed so dramatically that a complete
rewrite is by far the easiest way forward.  Rewriting also enables the
project to be relicensed.