Add initial barebones implementation
This implementation is essentially the WyilFilePrinter with some minor
rebranding.  For example, to output blocks using curly braces and to end
statements with semi-colons.

At the moment, the implementation is pretty rough and certainly does not
compile very much usefully.  However, it will generate a usable
implementation of a sum(int[]) function.  An important issue here,
however, is the issue of cloning (or the lack thereof) of mutable
data structures.