Add support for runtime type tests
This implements support for the runtime type test operator "is".  This
is implemented in a relatively simplistic fashion by simply traversing
the object in question trying to decide whether it's a match or not.

The mechanism could certainly be optimised a lot.  For example, when
looking at a union of records, which can try to decide the least number
of field tests needed to distinguish them, etc.

Finally, there are two difficult problems: references and
functions/methods.  Neither of these have an obvious approach for
implementing type tests.