Update representation of records + references
This updates the internal representation of records such that they now
employ the generic "Wy.Record" constructor.  This means we can now
distinguish Whiley records from "external" records using the JavaScript
"instanceof" operator.  This is not a bad solution, though it remains to
see whether we can do this for arrays as well.

Likewise, this also updates the representation of Whiley references to
use the "Wy.Ref" constructor.  This means that Whiley references are
"boxed".

At this stage, there is still a problem with external references as
these have a different representation from Whiley references.  It's not
completely clear to me how to resolve this.