Add support for bitwise operators + quantifiers
This fixes support for bitwise operators which were generating values
larger than a byte.  For example, the leftshift operator was not
clipping at 8bits.  This is resolved simply using "& 0xff", etc.

In addition, this adds support for quantifiers which are translated in a
functional style where a lambda function is mapped over the given array
range.  This is a reasonably effective solution which retains a similar
look to the original.