Add Type Testing for bytes, functions and refs
This adds runtime type testing support for bytes, functions and
references.  For bytes, the type test simply checks whether or not its
an integer (hence, this is unsound).  Likewise, for functions, this
simply checks whether it's any kind of function (i.e. because there is
not good embedded information).  Finally, references work though
presumably the general approach could fail in the presence of cycles.

Overall, the question of how to implement functions looms large.
Reaslistically, if it is possible to implement type testing using type
tags, then this would most likely be a better approach.