Updated DOM example.
This updates the DOM example to be a little more HTML compliant.  In
particular, this is now the beginning of a proper HTML library.  It's
not clear exactly where this should reside, but it probably should go
somewhere in the standard library.

This also adds the necessary function Wy.fromString() which converts a
Whiley ASCII string into a JavaScript string.  We need to add the
corresponding method to go the other way, of course.