Implement type tests
This implements type tests using existing JavaScript runtime type information.
The implementation is quite clever in optimising the generated code as much as
possible.  In particular, it uses the new StrictIncompleteSubtypeOperator which
has worked out very well.

However, it is clear now that using the existing runtime type information is not
sufficient for a general solution.  In fact, we must adopt a tagged union
approach.  Therefore, eventually, this will be replaced with an alternative
mechanism.