Support for ECMAScript Standards (ES5/ES6) #27
This adds support for compiling to different JavaScript platforms (e.g. ES5
versus ES6).  For the moment, ES5 is required for running under Nashhorn.
However, ES6 is set as the default since most browers support this.

In addition, a target ES6_BIGINT is supported for platforms which support the
new TC39 BigInt proposal.  However, this is not working properly yet for reasons
unknown.