Support building dependencies #35
This adds support for compiling package dependencies into JavaScript.
These are dumped into the same place as the main js file (e.g. bin/).
They are given the name of the package in question.  They build using
the configuration options of the current package, not any options
specified in the config of the dependency.