Support Bidirectional Type Inference
This updates the code base to be compatible with WyC v0.7.0.  This includes
updated the test cases as necessary, (including updating those which previously
lacked either the public of export modifier).

This also adds support for including the return type of a function or method in
the mangle.  This is necessary with the new type inference algorithm as this
can actually distinguish methods based solely on their return type.

One outcome of this, is that the AbstractTranslator is now significantly
simplified.  Previously, it performed forward type propagation, but this is now
unnecessary.  However, it's likely to need further changes in the future.