Preliminary support for js::core::string
This puts in place preliminary support for this native datatype.  Specifically,
constant strings are now created without coercions when assigned to a native
string.  Unfortunately, variable loads from an ascii::string don't force a
coercion because the variable expression type is not set correctly (for reasons
unknown).

Defining the native string type as a constrained type offers some benefits in
terms of reducing the cases for implicit coercions and the like.